requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
cryptography==41.0.3
dataclasses-json==0.5.14
dnspython==2.4.2
docutils==0.20.1
email-validator==2.0.0.post2
emoji==2.8.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.103.0
fastapi-cloud-tasks @ git+https://github.com/johncarpenter/fastapi-cloud-tasks.git@8ad3b41e29985bb69482564ebcb962698c6eaac8
filelock==3.0.12
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.9.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-scheduler==2.11.1
google-cloud-tasks==2.14.1
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.8
iniconfig==2.0.0
joblib==1.3.2
langchain==0.0.279
langsmith==0.0.33
lockfile==0.12.2
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pgvector==0.2.1
Pillow==10.0.0
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.0
pytest-env==1.0.1
python-daemon==2.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.26.0
rsa==4.9
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.12
urllib3==1.25.11
uvicorn==0.23.2
yarl==1.9.2
